+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 
-## /10/2023 - v4.0.3
+## 09/11/2023 - v4.0.4
+
+- **FileFormat Anycubic:** 
+  - (Fix) Preview marker size and `PropertyFields` for given file version
+  - (Fix) Do not cache and reuse layer images if equal on file write, as that is causing "file damage" message (#782, #787)
+- **Tools:**
+  - **Mask:**
+    - (Improvement) Apply the mask with Multiply instead of BitwiseAnd (#747)
+  - **Layer and pixel arithmetic**:
+    - (Improvement) Use the byte scale value (1/255) for multiply operations
+  - **Infill:**
+    - (Add) Floor/Ceil thickness (#785)
+    - (Fix) Remove the debug blocking window when selecting the honeycomb pattern
+- (Upgrade) .NET from 6.0.23 to 6.0.24
+
+## 19/10/2023 - v4.0.3
 
 - (Change) macOS: Change from Control to Command key to activate shortcuts from main menu (#766)
 - (Upgrade) .NET from 6.0.22 to 6.0.23