+++ RELEASE_NOTES.md
@@ -1,4 +1,13 @@
-- (Change) macOS: Change from Control to Command key to activate shortcuts from main menu (#766)
-- (Upgrade) .NET from 6.0.22 to 6.0.23
-- (Upgrade) AvaloniaUI from 11.0.4 to 11.0.5
+- **FileFormat Anycubic:**
+  - (Fix) Preview marker size and `PropertyFields` for given file version
+  - (Fix) Do not cache and reuse layer images if equal on file write, as that is causing "file damage" message (#782, #787)
+- **Tools:**
+  - **Mask:**
+    - (Improvement) Apply the mask with Multiply instead of BitwiseAnd (#747)
+  - **Layer and pixel arithmetic**:
+    - (Improvement) Use the byte scale value (1/255) for multiply operations
+  - **Infill:**
+    - (Add) Floor/Ceil thickness (#785)
+    - (Fix) Remove the debug blocking window when selecting the honeycomb pattern
+- (Upgrade) .NET from 6.0.23 to 6.0.24